GraphRAG/requirements.txt
streamlit
faiss-cpu
spacy
networkx
python-dotenv
sentence-transformers
langchain
langchain-openai
PyMuPDF
python-docx